feat: Add wiki introduction
